--- src/Core/tests/DeviceTests/Handlers/Button/ButtonFocusTests.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.Maui.DeviceTests.Stubs;
-
-namespace Microsoft.Maui.DeviceTests
-{
-#if WINDOWS
-	[Category(TestCategory.Button)]
-	public class ButtonFocusTests : FocusHandlerTests<ButtonHandler, ButtonStub, VerticalStackLayoutStub>
-	{
-		public ButtonFocusTests()
-		{
-		}
-	}
-#endif
-}
