--- src/Core/tests/DeviceTests/Handlers/Entry/EntryFocusTests.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.Maui.DeviceTests.Stubs;
-
-namespace Microsoft.Maui.DeviceTests
-{
-	[Category(TestCategory.Entry)]
-	public class EntryFocusTests : FocusHandlerTests<EntryHandler, EntryStub, VerticalStackLayoutStub>
-	{
-		public EntryFocusTests()
-		{
-		}
-	}
-}
