+++ src/Core/tests/DeviceTests/Handlers/Entry/EntryHandlerTests.cs
@@ -614,5 +614,23 @@ namespace Microsoft.Maui.DeviceTests
 		public class EntryTextStyleTests : TextStyleHandlerTests<EntryHandler, EntryStub>
 		{
 		}
+
+		[Category(TestCategory.Entry)]
+		public class EntryFocusTests : FocusHandlerTests<EntryHandler, EntryStub, VerticalStackLayoutStub>
+		{
+		}
+
+		[Category(TestCategory.Entry)]
+		public class EntryTextInputTests : TextInputHandlerTests<EntryHandler, EntryStub>
+		{
+			protected override void SetNativeText(EntryHandler entryHandler, string text) =>
+				EntryHandlerTests.SetNativeText(entryHandler, text);
+
+			protected override int GetCursorStartPosition(EntryHandler entryHandler) =>
+				EntryHandlerTests.GetCursorStartPosition(entryHandler);
+
+			protected override void UpdateCursorStartPosition(EntryHandler entryHandler, int position) =>
+				EntryHandlerTests.UpdateCursorStartPosition(entryHandler, position);
+		}
 	}
 }
