--- src/Core/tests/DeviceTests/Handlers/Entry/EntryTextInputTests.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.Maui.DeviceTests.Stubs;
-
-namespace Microsoft.Maui.DeviceTests
-{
-	[Category(TestCategory.Entry)]
-	public class EntryTextInputTests : TextInputHandlerTests<EntryHandler, EntryStub>
-	{
-		protected override void SetNativeText(EntryHandler entryHandler, string text)
-		{
-			EntryHandlerTests.SetNativeText(entryHandler, text);
-		}
-		protected override int GetCursorStartPosition(EntryHandler entryHandler)
-		{
-			return EntryHandlerTests.GetCursorStartPosition(entryHandler);
-		}
-
-		protected override void UpdateCursorStartPosition(EntryHandler entryHandler, int position)
-		{
-			EntryHandlerTests.UpdateCursorStartPosition(entryHandler, position);
-		}
-	}
-}
