+++ src/Core/tests/DeviceTests/Handlers/SearchBar/SearchBarHandlerTests.cs
@@ -303,5 +303,29 @@ namespace Microsoft.Maui.DeviceTests
 		public class SearchBarTextStyleTests : TextStyleHandlerTests<SearchBarHandler, SearchBarStub>
 		{
 		}
+
+		[Category(TestCategory.SearchBar)]
+		public class SearchBarTextInputTests : TextInputHandlerTests<SearchBarHandler, SearchBarStub>
+		{
+			protected override void SetNativeText(SearchBarHandler searchBarHandler, string text) =>
+				SearchBarHandlerTests.SetNativeText(searchBarHandler, text);
+
+			protected override int GetCursorStartPosition(SearchBarHandler searchBarHandler) =>
+				SearchBarHandlerTests.GetCursorStartPosition(searchBarHandler);
+
+			protected override void UpdateCursorStartPosition(SearchBarHandler searchBarHandler, int position) =>
+				SearchBarHandlerTests.UpdateCursorStartPosition(searchBarHandler, position);
+		}
+
+		// TODO: only iOS is working with the search bar focus tests
+#if IOS || MACCATALYST
+		[Category(TestCategory.SearchBar)]
+		public class SearchBarFocusTests : FocusHandlerTests<SearchBarHandler, SearchBarStub, VerticalStackLayoutStub>
+		{
+			public SearchBarFocusTests()
+			{
+			}
+		}
+#endif
 	}
 }
